MVP ARchitecture in ChrgesDialogFragment
