+++ mifosng-android/src/main/java/com/mifos/mifosxdroid/dialogfragments/chargedialog/ChargeDialogPresenter.java
@@ -1,0 +1,94 @@
+package com.mifos.mifosxdroid.dialogfragments.chargedialog;
+
+import com.mifos.api.DataManager;
+import com.mifos.mifosxdroid.base.BasePresenter;
+import com.mifos.objects.accounts.savings.Charge;
+import com.mifos.objects.client.Charges;
+import com.mifos.services.data.ChargesPayload;
+
+import javax.inject.Inject;
+
+import retrofit.client.Response;
+import rx.Subscriber;
+import rx.Subscription;
+import rx.android.schedulers.AndroidSchedulers;
+import rx.schedulers.Schedulers;
+
+/**
+ * Created by Rajan Maurya on 08/06/16.
+ */
+public class ChargeDialogPresenter extends BasePresenter<ChargeDialogMvpView> {
+
+    private final DataManager mDataManager;
+    private Subscription mSubscription;
+
+    @Inject
+    public ChargeDialogPresenter(DataManager dataManager) {
+        mDataManager = dataManager;
+    }
+
+    @Override
+    public void attachView(ChargeDialogMvpView mvpView) {
+        super.attachView(mvpView);
+    }
+
+    @Override
+    public void detachView() {
+        super.detachView();
+        if (mSubscription != null) mSubscription.unsubscribe();
+    }
+
+    public void loadAllChargesV2(int clientId) {
+        checkViewAttached();
+        getMvpView().showProgressbar(true);
+        if (mSubscription != null) mSubscription.unsubscribe();
+        mSubscription = mDataManager.getAllChargesV2(clientId)
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribeOn(Schedulers.io())
+                .subscribe(new Subscriber<Response>() {
+                    @Override
+                    public void onCompleted() {
+                        getMvpView().showProgressbar(false);
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        getMvpView().showProgressbar(false);
+                        getMvpView().showFetchingError("Failed to fetch Charges");
+                    }
+
+                    @Override
+                    public void onNext(Response response) {
+                        getMvpView().showProgressbar(false);
+                        getMvpView().showAllChargesV2(response);
+                    }
+                });
+    }
+
+    public void createCharges(int clientId, ChargesPayload payload){
+        checkViewAttached();
+        getMvpView().showProgressbar(true);
+        if (mSubscription != null) mSubscription.unsubscribe();
+        mSubscription = mDataManager.createCharges(clientId, payload)
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribeOn(Schedulers.io())
+                .subscribe(new Subscriber<Charges>() {
+                    @Override
+                    public void onCompleted() {
+                        getMvpView().showProgressbar(false);
+                    }
+
+                    @Override
+                    public void onError(Throwable e) {
+                        getMvpView().showProgressbar(false);
+                        getMvpView().showFetchingError("Try Again");
+                    }
+
+                    @Override
+                    public void onNext(Charges charges) {
+                        getMvpView().showProgressbar(false);
+                        getMvpView().showChargesCreatedSuccessfully(charges);
+                    }
+                });
+    }
+}
